totalcommander: fix uninstallation on non-English OSes
The standard system message box button texts are localized, e.g. on Polish
Windows 10 "Yes" is "Tak". This causes the ahk script to fail and leaves
the message box warning about removing the 32-bit Total Commander open.

The third argument to ControlClick is interpreted in the same way as in
WinWait - any text fragment from the window containing the control, not
neccessarily the text of the control to click. So, let's use the same text
as for WinWait.