shimgen helper
shimgen nicely goes through all the executables but fails to identitfy them as "gui".  It would be nice not to be stuck with a non fonctionning command prompt everytime I fire up one of nir's utilities.  This adds an empty ".gui" file to every in tools and tools\64 save for "nircmd.exe" and "nircmdc.exe" which are "console" programs.