#229 firefox-nightly (doesn't work)
Did a bunch of work to make firefox-nightly work, but it will not work because
nuspec version element is field limited to 10 chars. I intended to use 12.
YYYYMMDD = 8
YYYYMMDDhhmm = 12
I only had one field to work with because I decided to redesignate a field with
alpha version. So 50.0-alpha1 would be 50.0.1-alpha
Since that won't work with the current limitation, I will have to either

* Only use up to the hour (YYYYMMDDHH) which should be accurate enough most of
  the time, but will break the package if an update is packaged in the same
  hour. However, since I am not scanning for updated versions that often, this
  is a non-issue.
  50.0.1.YYMMDDHH-alpha
* Do not replace the digit appended to "alpha". Mozilla seems to always use this
  digit, but if they decide not to, this update routine would be broke anyway
  because no digit means 5.0..YYMMDDHH-alpha so that would need to be fixed if
  the former is chosen. This option would prevent having to make that fix.
  If I keep as "alpha1", then I have 2 fields to work with so I can split up
  the date string.
  50.0.YYYYMMDD.hhmm-alpha1

I plan on making the simpler tweak (at this point) of just omitting the minutes